+++ eic.h
@@ -7,21 +7,21 @@
 struct zmcu_eic {
 
 	/* 0x00: Control */
-	uint32_t volatile CTRL;
+	uint8_t volatile CTRL;
 #define EIC_CTRL_ENABLE				(1u << 1)
 #define EIC_CTRL_SWRST				(1u << 0)
 
 	/* 0x01: Status */
-	uint32_t volatile STATUS;
+	uint8_t volatile STATUS;
 #define EIC_STATUS_SYNCBUSY			(1u << 7)
 
 	/* 0x02: Non-Maskable Interrupt Control */
-	uint32_t volatile NMICTRL;
+	uint8_t volatile NMICTRL;
 #define EIC_NMICTRL_NMIFILTEN			(1u << 3)
 #define EIC_NMICTRL_NMISENSE(x)			((x) << 0)
 
 	/* 0x03: Non-Maskable Interrupt Flag Status and Clear */
-	uint32_t volatile NMIFLAG;
+	uint8_t volatile NMIFLAG;
 #define EIC_NMIFLAG_NMI				(1u << 0)
 
 	/* 0x04: Event Control */
